Update README.md
Car Insurance 
This project is an MVC web application that mimics a car insurance website. It takes the user's input on policy information and calculates a quote based on business logic. There is also an admin page that allows staff to view all of the quotes that have been issued.